requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-dotenv==1.4.2
django-ranged-response==0.2.0
django-recaptcha2==1.4.1
django-simple-captcha==0.6.0
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
